// src/components/formSignUp/formSignUp.types.ts
export interface IFormSignUp {
  name: string;
  lastName: string;
  birthDay?: string;
  gender?: string;
  ocupation?: string;
  email: string;
  emailSecondary?: string;
  cellPhone?: number;
  phone?: number;
  state?: string;
  city?: string;
  address?: string;
  postal?: string;
  creditID: string;
  password: string;
}